feat: agregar stock de un producto

+++ servicio-lote/src/main/java/com/proyecto/servicio_lote/domain/repositories/LoteRepository.java
@@ -3,5 +3,5 @@ package com.proyecto.servicio_lote.domain.repositories;
 import com.proyecto.servicio_lote.domain.models.Lote;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoteRepository extends JpaRepository<Lote, Integer> {
+public interface LoteRepository extends JpaRepository<Lote, Long> {
 }
