+++ servicio-lote/src/main/java/com/proyecto/servicio_lote/domain/repositories/ProveedorRepository.java
@@ -3,5 +3,5 @@ package com.proyecto.servicio_lote.domain.repositories;
 import com.proyecto.servicio_lote.domain.models.Proveedor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProveedorRepository extends JpaRepository<Proveedor, Integer> {
+public interface ProveedorRepository extends JpaRepository<Proveedor, Long> {
 }
